feat(ui): 优化全屏播放音量控制交互与视觉样式，统一搜索界面图标

+++ src/data/config.rs
@@ -40,7 +40,7 @@ impl GraphicsProtocol {
             GraphicsProtocol::Off => 0,
             GraphicsProtocol::Halfblocks => 1,
         };
-        let next = (current as i32 + delta).rem_euclid(Self::ALL.len() as i32) as usize;
+        let next = (current + delta).rem_euclid(Self::ALL.len() as i32) as usize;
         Self::ALL[next]
     }
